Add alert on user creation
